airflow-dag-sample/requirements.txt
git+http://github.com/graphgrid/graphgrid-sdk-python@2.0-nmt-train-pipeline#egg=graphgrid-sdk
airflow-provider-graphgrid==2.0.0

apache-airflow==2.2.2

apache-airflow-providers-docker==2.3.0
requests==2.27.1
javaproperties==0.8.1
neo4j==1.7.6
neobolt==1.7.17
neotime==1.7.4
pytz==2021.1
six==1.15.0
boto3==1.17.27
fire==0.4.0
airflow-provider-graphgrid==2.0.0